workflow_police/requirements.txt
jira>=2.0.0
pytest>=3.5.0
pyaml>=20.4.0
gitpython>=3.1.7
graypy>=2.1.0
